Feature/conf command

+++ ARCaDIUS_Demo/Shutter.h
@@ -43,7 +43,7 @@ class Shutter{
         // @param control: Set 0 to use goto_open, Set 1 to use goto_controlled
 
 
-        void timedExposure(state pos, unsigned int exposure_time, bool controller);
+        void timedExposure(state pos, unsigned long int exposure_time, bool controller);
 
 
 };
